Moved the shared parse-excel test helper back under packages/excel-parser/src/__tests__/helpers/parse-excel.ts and tightened packages/excel-parser/jest.config.ts so only real *.test.ts / *.spec.ts files under __tests__ are discovered by Jest. This keeps helpers colocated with tests without executing them as suites. Also fixed temp workbook filename collisions in the shared helper by switching to randomUUID()-based filenames, which resolved intermittent failures in the full npm run test:unit --workspace=@supplier-config/excel-parser run.
