+++ packages/excel-parser/jest.config.ts
@@ -13,7 +13,7 @@ const config: Config = {
   coverageDirectory: "./.reports/unit/coverage",
 
   // Indicates which provider should be used to instrument code for coverage
-  coverageProvider: "v8",
+  coverageProvider: "babel",
 
   coverageThreshold: {
     global: {
@@ -48,7 +48,7 @@ const config: Config = {
 
   testEnvironment: "node",
   testPathIgnorePatterns: ["/node_modules/", "/dist/"],
-  testMatch: ["**/__tests__/**/*.ts", "**/?(*.)+(spec|test).ts"],
+  testMatch: ["**/__tests__/**/*.test.ts", "**/__tests__/**/*.spec.ts"],
   moduleFileExtensions: ["ts", "js", "json", "node"],
 
   transform: {
